requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
boto3==1.34.0
botocore==1.34.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
distro==1.8.0
Django==4.2.6
django-crispy-forms==2.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.3
jmespath==1.0.1
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
openai==1.5.0
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.10.3
requests==2.28.1
s3transfer==0.9.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tellopy==0.6.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.10
